Adds functionality to operator buttons
Currently, this is a very basic implementation. It expects the first number to be entered, followed by the operator, followed by the second number and then the equals button. This will be improved in future iterations. A clear button will be implemented some time in the future, along with a decimal button and other features.